Some refactoring so node bucket is more usable
